Moved utility classes to util package
