+++ app/src/main/java/me/proteus/myeye/visiontests/CircleTest.kt
@@ -29,12 +29,11 @@ import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import me.proteus.myeye.GrammarType
 import me.proteus.myeye.R
-import me.proteus.myeye.ScreenScalingUtils.getScreenInfo
+import me.proteus.myeye.util.ScreenScalingUtils.getScreenInfo
 import me.proteus.myeye.SerializableStage
 import me.proteus.myeye.TestResult
 import me.proteus.myeye.VisionTest
